Fix syntax, align sensitivity logic, and implement frontend auth routing.
- Fixed misplaced import in `src/ui/Login.jsx`.
- Corrected inverted sensitivity threshold logic in `src/nlp/cd-detector.js`.
- Enabled JWT token interceptor in `src/services/api.js`.
- Refactored `src/ui/App.jsx` to implement `BrowserRouter` and `ProtectedRoute`.
- Extracted main dashboard UI to `src/ui/Dashboard.jsx`.
